Turned notepad/input into fixed 32-bit integers.
This will add some consistency to the interpreter, including with how overflows are handled (hopefully). To be honest, I'd prefer 64-bit because there aren't a lot of variables and 64-bit is large enough to handle almost every use, but 32-bit is more widely supported than 64-bit.
Please note that this is untested. I don't have a C compiler on this computer, so I'd have to remote into a terminal, clone the repo, make the change, then test it there (I'm too lazy).